+++ src/preprocessing/pattern/MultiGraphPatternPreprocessor.cpp
@@ -15,7 +15,7 @@ namespace sgf
 
 MultiGraphPatternPreprocessor::MultiGraphPatternPreprocessor(
     std::vector<ColoredGraph>& graph_library, const bool is_directed, const uint32_t pattern_number,
-    const uint32_t alive_precent, LoggerHandler logger)
+    const double alive_precent, LoggerHandler logger)
     : m_graph_library(graph_library)
     , M_IS_DIRECTED(is_directed)
     , M_PATTERN_NUMBER(pattern_number)
